+++ jingle/src/context.rs
@@ -1,96 +1,91 @@
 use crate::modeling::State;
-use jingle_sleigh::{ArchInfoProvider, SpaceInfo, VarNode};
+use jingle_sleigh::{ArchInfoProvider, SleighArchInfo, SpaceInfo, VarNode};
 use std::ops::Deref;
 use std::rc::Rc;
-use z3::Context;
+use z3::{Context, Translate};
 
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct CachedArchInfo {
-    registers: Vec<(VarNode, String)>,
-    spaces: Vec<SpaceInfo>,
-    default_code_space: usize,
-}
-
-impl ArchInfoProvider for JingleContext<'_> {
+impl ArchInfoProvider for JingleContext {
     fn get_space_info(&self, idx: usize) -> Option<&SpaceInfo> {
-        self.info.spaces.get(idx)
+        self.info.get_space_info(idx)
     }
 
     fn get_all_space_info(&self) -> impl Iterator<Item = &SpaceInfo> {
-        self.info.spaces.iter()
+        self.info.get_all_space_info()
     }
 
     fn get_code_space_idx(&self) -> usize {
-        self.info.default_code_space
+        self.info.get_code_space_idx()
     }
 
     fn get_register(&self, name: &str) -> Option<&VarNode> {
-        self.info
-            .registers
-            .iter()
-            .find(|(_, reg_name)| reg_name.as_str() == name)
-            .map(|(vn, _)| vn)
+        self.info.get_register(name)
     }
 
     fn get_register_name(&self, location: &VarNode) -> Option<&str> {
-        self.info
-            .registers
-            .iter()
-            .find(|(vn, _)| vn == location)
-            .map(|(_, name)| name.as_str())
+        self.info.get_register_name(location)
     }
 
     fn get_registers(&self) -> impl Iterator<Item = (&VarNode, &str)> {
-        self.info.registers.iter().map(|(a, b)| (a, b.as_str()))
+        self.info.get_registers()
     }
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct JingleContextInternal<'ctx> {
-    pub z3: &'ctx Context,
-    pub info: CachedArchInfo,
+pub struct JingleContextInternal {
+    pub z3: Context,
+    pub info: SleighArchInfo,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub struct JingleContext<'ctx>(Rc<JingleContextInternal<'ctx>>);
+pub struct JingleContext(Rc<JingleContextInternal>);
 
-impl JingleContext<'_> {
-    pub(crate) fn translate<'a>(&self, ctx: &'a Context) -> JingleContext<'a> {
+impl JingleContext {
+    pub(crate) fn translate(&self, ctx: &Context) -> JingleContext {
         JingleContext(Rc::new(JingleContextInternal {
-            z3: ctx,
+            z3: ctx.clone(),
             info: self.info.clone(),
         }))
     }
 }
 
-impl<'ctx> Deref for JingleContext<'ctx> {
-    type Target = JingleContextInternal<'ctx>;
+impl Deref for JingleContext {
+    type Target = JingleContextInternal;
 
     fn deref(&self) -> &Self::Target {
         self.0.as_ref()
     }
 }
-impl<'ctx> JingleContext<'ctx> {
-    pub fn new<S: ArchInfoProvider>(z3: &'ctx Context, r: &S) -> Self {
+impl JingleContext {
+    pub fn new<S: ArchInfoProvider>(z3: &Context, r: &S) -> Self {
         Self(Rc::new(JingleContextInternal {
-            z3,
-            info: CachedArchInfo {
-                spaces: r.get_all_space_info().cloned().collect(),
-                registers: r
-                    .get_registers()
-                    .map(|(a, b)| (a.clone(), b.to_string()))
-                    .collect(),
-                default_code_space: r.get_code_space_idx(),
-            },
+            z3: z3.clone(),
+            info: SleighArchInfo::new(
+                r.get_registers(),
+                r.get_all_space_info(),
+                r.get_code_space_idx(),
+            ),
         }))
     }
-    pub fn fresh_state(&self) -> State<'ctx> {
+
+    pub fn ctx(&self) -> &Context {
+        &self.z3
+    }
+    pub fn fresh_state(&self) -> State {
         State::new(self)
     }
 
-    pub fn with_fresh_z3_context(&self, z3: &'ctx Context) -> Self {
+    pub fn with_fresh_z3_context(&self, z3: &Context) -> Self {
         Self(Rc::new(JingleContextInternal {
-            z3,
+            z3: z3.clone(),
+            info: self.info.clone(),
+        }))
+    }
+}
+
+unsafe impl Translate for JingleContext {
+    fn translate(&self, dest: &Context) -> Self {
+        Self(Rc::new(JingleContextInternal {
+            z3: dest.clone(),
             info: self.info.clone(),
         }))
     }