+++ jingle/src/python/modeled_block.rs
@@ -1,4 +1,5 @@
 use crate::JingleContext;
+use crate::display::JingleDisplayable;
 use crate::modeling::{ModeledBlock, ModelingContext};
 use crate::python::instruction::PythonInstruction;
 use crate::python::resolved_varnode::PythonResolvedVarNode;
@@ -9,12 +10,12 @@ use pyo3::{PyResult, pyclass, pymethods};
 
 #[pyclass(unsendable, name = "ModeledBlock")]
 pub struct PythonModeledBlock {
-    pub instr: ModeledBlock<'static>,
+    pub instr: ModeledBlock,
 }
 
 impl PythonModeledBlock {
     pub fn new<T: Iterator<Item = Instruction>>(
-        jingle: &JingleContext<'static>,
+        jingle: &JingleContext,
         i: T,
     ) -> PyResult<PythonModeledBlock> {
         Ok(Self {
@@ -51,29 +52,29 @@ impl PythonModeledBlock {
     /// for only those representing actual locations in processor memory:
     /// constants and "internal" varnodes are filtered out
     pub fn get_input_vns(&self) -> PyResult<VarNodeIterator> {
-        let s = self.instr.get_original_state().clone();
-        let filtered: Result<Vec<PythonResolvedVarNode>, _> = self
+        let info = self.instr.get_jingle().info.clone();
+        let filtered: Vec<PythonResolvedVarNode> = self
             .instr
             .get_inputs()
             .into_iter()
-            .map(|g| g.display(&s).map(PythonResolvedVarNode::from))
+            .map(|g| PythonResolvedVarNode::from(g.display(&info)))
             .collect();
-        let filtered = filtered?.into_iter();
-        Ok(VarNodeIterator::new(filtered.into_iter()))
+        let filtered = filtered.into_iter();
+        Ok(VarNodeIterator::new(filtered))
     }
 
     /// A list of the output varnodes to the block, filtering
     /// for only those representing actual locations in processor memory:
     /// "internal" varnodes are filtered out
     pub fn get_output_vns(&self) -> PyResult<VarNodeIterator> {
-        let s = self.instr.get_final_state().clone();
-        let filtered: Result<Vec<PythonResolvedVarNode>, _> = self
+        let s = self.instr.get_jingle().info.clone();
+        let filtered: Vec<PythonResolvedVarNode> = self
             .instr
             .get_outputs()
             .into_iter()
-            .map(|g| g.display(&s).map(PythonResolvedVarNode::from))
+            .map(|g| PythonResolvedVarNode::from(g.display(&s)))
             .collect();
-        let filtered = filtered?.into_iter();
-        Ok(VarNodeIterator::new(filtered.into_iter()))
+        let filtered = filtered.into_iter();
+        Ok(VarNodeIterator::new(filtered))
     }
 }