+++ jingle/src/python/z3/bitvec.rs
@@ -6,7 +6,7 @@ use pyo3::{IntoPyObject, IntoPyObjectExt, Py, PyAny, PyResult, Python};
 use z3::ast::{Ast, BV};
 use z3_sys::Z3_ast;
 
-impl TryIntoPythonZ3 for BV<'static> {
+impl TryIntoPythonZ3 for BV {
     fn try_into_python(self) -> PyResult<Py<PyAny>> {
         Python::with_gil(|py: Python| {
             let z3_mod = PyModule::import(py, "z3")?;
@@ -22,7 +22,7 @@ impl TryIntoPythonZ3 for BV<'static> {
     }
 }
 
-impl TryFromPythonZ3 for BV<'static> {
+impl TryFromPythonZ3 for BV {
     fn try_from_python(py_bv: Py<PyAny>) -> PyResult<Self> {
         Python::with_gil(|py| {
             let z3 = get_python_z3()?;