+++ jingle/CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2](https://github.com/toolCHAINZ/jingle/compare/jingle-v0.2.1...jingle-v0.2.2) - 2025-08-12
+
+### Fixed
+
+- consolidate one-off display types ([#81](https://github.com/toolCHAINZ/jingle/pull/81))
+
+## [0.2.1](https://github.com/toolCHAINZ/jingle/compare/jingle-v0.2.0...jingle-v0.2.1) - 2025-08-11
+
+### Other
+
+- update z3 ([#79](https://github.com/toolCHAINZ/jingle/pull/79))
+
+## [0.2.0](https://github.com/toolCHAINZ/jingle/compare/jingle-v0.1.4...jingle-v0.2.0) - 2025-08-06
+
+### Added
+
+- back edge analysis ([#76](https://github.com/toolCHAINZ/jingle/pull/76))
+- add basic analysis ([#74](https://github.com/toolCHAINZ/jingle/pull/74))
+
+### Other
+
+- [**breaking**] bump z3 ([#78](https://github.com/toolCHAINZ/jingle/pull/78))
+
 ## [0.1.4](https://github.com/toolCHAINZ/jingle/compare/jingle-v0.1.3...jingle-v0.1.4) - 2025-07-16
 
 ### Fixed